requirements.txt
pandas>=1.3.0
numpy>=1.20.0
requests>=2.25.0
ttkthemes>=3.2.0
packaging>=21.0
pyperclip>=1.8.0
plotly>=5.3.0
openpyxl>=3.0.0  # 用於 Excel 匯出功能
matplotlib==3.7.1
scikit-learn==1.3.0
seaborn==0.12.2
kaleido==0.2.1
win10toast-click>=0.1.2  # Windows 10 通知支援
plotly-express>=0.4.0  # 用於進階視覺化 